+++ learn_kernels.cuh
@@ -7,6 +7,6 @@
 #define NUM_SPARSE_GRAD_BLOCKS               4096
 #define NUM_SPARSE_GRAD_THREADS_PER_BLOCK    512
 
-__global__ void kMultiplyBySigmoidGrad(float* act, float* target, const unsigned int len);
+__global__ void kMultiplyBySigmoidGrad(double* act, double* target, const unsigned int len);
 
 #endif
